+++ packages/autofe-scripts/gulpfile.js/tasks/build.js
@@ -3,11 +3,6 @@
 const gulp = require('gulp');
 const runSequence = require('run-sequence');
 
-gulp.task('build', (cb) => {
-  runSequence(
-    ['clean'],
-    ['copy', 'fonts', 'svg', 'images'],
-    ['sass', 'js', 'webpack', 'html', 'markdown'],
-    ['html-bundle'],
-    cb);
+gulp.task('build', cb => {
+  runSequence(['clean'], ['sprite'], ['copy', 'fonts', 'svg', 'images'], ['sass', 'js', 'webpack', 'html', 'markdown'], ['html-bundle'], cb);
 });