+++ packages/autofe-scripts/gulpfile.js/tasks/default.js
@@ -3,12 +3,6 @@
 const gulp = require('gulp');
 const runSequence = require('run-sequence');
 
-gulp.task('default', (cb) => {
-  runSequence(
-    ['clean'],
-    ['copy', 'fonts', 'svg', 'images'],
-    ['sass', 'js', 'webpack', 'html', 'markdown'],
-    ['watch'],
-    ['browserSync'],
-    cb);
+gulp.task('default', cb => {
+  runSequence(['clean'], ['sprite'], ['copy', 'fonts', 'svg', 'images'], ['sass', 'js', 'webpack', 'html', 'markdown'], ['watch'], ['browserSync'], cb);
 });